vectra/requirements.txt
requests==2.27.1
azure-identity==1.10.0
pytest==7.2.0